web app/requirements.txt
Flask
google-cloud-firestore